requirements.txt
aiohttp==3.6.2
python-dotenv==0.10.3
aiofiles==0.4.0
aiohttp-jinja2==1.2.0